[FIX] Fixes the correct simulation and materialisation of deletions.
Following fixes are applied in this commit:
* No small variant can be simulated that overlaps with a deletion or any other structural variant.
* The deletion end position is fixed, as well its translation into a journal.

In general, the code could be refactored a lot to reduce unnecessary copies of the same code.
In many places this led to some errors since the deletions are represented with negative sizes and always need to be put carefully into the correct context.